requirements.txt
torch
torchaudio
transformers
librosa
numpy
pandas
scipy
umap-learn
scikit-learn
matplotlib
plotly
click
tqdm
soundfile
requests
streamlit>=1.28.0

ipykernel
ipython
nbformat>=4.2.0

# Storage backends (optional)
python-dotenv>=1.0.0

# Astra DB (required for astra backend)
astrapy>=0.7.0